requirements.txt
aiogram
openai
googletrans==4.0.0rc1
apscheduler
python-dotenv
sqlite-utils 